GUI sketching cont'd.
- Sketched GUI widgets
- Added new stock fonts (dejavu)
- moved old default font (Courier New) to obsolete
- Added node-specific material modifier support for shader
- Changed many ulong's to size_t's (for 32-bit support)

+++ engine/render/layer.d
@@ -97,7 +97,7 @@ class BasicNodeGroup : CollectableNodeGroup
 {
     bool[Node] nodes;
 
-    ulong length() { return nodes.length; }
+    size_t length() { return nodes.length; }
 
     this() { }
